+++ src/storage/hazelcast.ts
@@ -1,0 +1,108 @@
+import { IMap } from "hazelcast-client";
+import { SignManager } from "../sign";
+import { SIGNATURE_SEPARATOR, StorageWrapper, deepFreeze } from ".";
+import { ICacheStorage } from "./storage-interface";
+
+/**
+ * Hazelcast cache storage (implements cache signatures)
+ * All items are stored in Hazelcast distributed map and are immutable after retrieval
+ * because they might be shared between threads to prevent data corruption
+ */
+export class HazelcastStorage implements ICacheStorage {
+  private cache: IMap<string, string>;
+
+  /**
+   * Create a wrapper class for Hazelcast IMap
+   * @param hazelcastMap Hazelcast IMap instance
+   */
+  constructor(hazelcastMap: IMap<string, string>) {
+    this.cache = hazelcastMap;
+  }
+
+  /**
+   * Retrieves cache key
+   * @param key
+   * @returns
+   */
+  async get<T>(key: string, immutable?: boolean): Promise<T | undefined> {
+    let cachedContent: string | null | undefined = await this.cache.get(key);
+
+    if (cachedContent) {
+      const signingKey = SignManager.obtainKey(key);
+
+      if (signingKey) {
+        const separatorPosition = cachedContent.indexOf(SIGNATURE_SEPARATOR);
+        const content = cachedContent.substring(separatorPosition + 1);
+        const signature = cachedContent.substring(0, separatorPosition);
+
+        if (!SignManager.verifySignedContent(content, signingKey, signature)) {
+          // treat as cache miss
+          return;
+        }
+
+        // remove signature from content
+        cachedContent = content;
+      }
+
+      try {
+        const wrapper: StorageWrapper<T> | undefined = (JSON.parse(
+          cachedContent
+        ) as unknown) as StorageWrapper<T>;
+
+        // this forces in memory object not to be changed between instances
+        if (immutable) {
+          deepFreeze(wrapper);
+        }
+
+        return wrapper?.value;
+      } catch (e) {
+        // treat JSON parse error as cache miss
+        return undefined;
+      }
+    }
+  }
+
+  /**
+   * Sets cache key
+   * @param key key to be set
+   * @param ttlMilliseconds milliseconds to keep in cache
+   * @param value value to store
+   */
+  async set<T>(
+    key: string,
+    ttlMilliseconds: number,
+    value: T,
+    immutable?: boolean
+  ): Promise<void> {
+    if (immutable) {
+      throw new Error(
+        "Immutability should be set during cache get for external caches"
+      );
+    }
+
+    const wrapper: StorageWrapper<T> = {
+      value,
+    };
+
+    const signingKey = SignManager.obtainKey(key);
+    let content = JSON.stringify(wrapper);
+
+    if (signingKey) {
+      const signature = SignManager.signContent(content, signingKey);
+      // append signature to content
+      content = signature + SIGNATURE_SEPARATOR + content;
+    }
+
+    // Hazelcast set with TTL
+    await this.cache.set(key, content, ttlMilliseconds);
+  }
+
+  /**
+   * Evict cache key
+   * @param key
+   * @returns Promise<void>
+   */
+  async evict(key: string): Promise<void> {
+    await this.cache.delete(key);
+  }
+}